♻️ [Refactoring]: 各ファイルでStyleインポートパスを新しいモデルに更新
- figma-types.tsからのStyle型インポートを削除
- src/models/style/style.jsからの新しいStyle型インポートに変更
- APIレスポンス型定義、データファイル、ツール型定義を更新
